Extend NEList and move utilities from Megaparsec
Problem: NEList.nonEmpty has a bug: it will never produce a '.some'.
We need this fixed for Megaparsec.
Additionally, there's some stuff defined in Megaparsec that would be
more useful in YatimaStdLib, and we want to move it here.

Solution: fixed the bug. Added functions 'void', 'between', and
'liftSeq2' to reasonable locations in YatimaStdLib.
Also, NEList was lacking some useful instances, e.g. of ToString,
so they were added along with Traversable and Pure instances.