Add Cached data type
Problem: there is no easy construct in Lean/YatimaStdLib that allows
to memoize and/or force just one value to be considered valid on the
type level.

Solution: added Cached data type, which allows to memoize a function
application and the resulting value. We also defined a Cached' utility
abbreviation that uses 'id' to allow for just the value memoization.