Add Subsingleton equality theorem for simp
Problem: Cached ultimately represents a singleton value, and so
two values of Cached of the same type must be equal to each other.
It'd be nice for the simplifier to be able to use that. We've had
a small theorem but forgot to put it in.

Solution: added the [simp] theorem back.