Add some RBMap utilities
Problem: we need to use RBMap in Megaparsec, simulating finite
computable sets through RBMap with Unit as nominal values. For ease
of use, we need a way to construct such a map from just a list of
keys, i.e. real values, as well as extract keys and merge and append
two such "sets".

Solution: added these functions to RBMap module. Also deleted the
Inhabited instance for RBMap as the same definition is present in
Lean's stdlib.