Better definition of `NEList` (#75)
Previously defined `NEList` was inefficient because conversions between
`NEList` and `List`, whichever way, were O(n). With this new definition,
conversions will be O(1). In fact, an `NEList` will have the same memory
layout as a `List`, so if the compiler is smart enough the conversion
will be 0 cost. There's also an added benefit of being able to reuse
functions on `List` when defining their `NEList` counterpart, leading to
more concise code.