Redefined NEList again (#88)
The `data/ne` version of `NEList` has a huge issue: it cannot be nested
inside another inductive. That's why I'm redefining it as a `head/tail`
list again. This change should not conflict with anything else, except
if it is using an explicit pattern match to extract either `data` or
`ne`